Revert "Revert "Unify code to use RequestServicesContainer always""
This reverts commit 61a670e0971fcf5e5c7a5d38353f460a27ccad4d.
